Fixed the Login and Register page and change the users database a bit
So I implemented the login/register page from laravel to autogenerate
views I change the database code to include id and other variables in
order to run it. so make sure to run the database again.
the css is stored in app.css if you want to change the layout. any
questions ask on facebook